Move Patroni to 'internal' & Cleanup
patroni.go has now been moved to 'internal/patroni/patroni.go' to align
with the new PostgreSQL Operator directory structure.  This commit also
includes some general cleanup per PR feedback.
