+++ apiserver/reloadservice/reloadimpl.go
@@ -24,7 +24,7 @@ import (
 	msgs "github.com/crunchydata/postgres-operator/apiservermsgs"
 	"github.com/crunchydata/postgres-operator/config"
 	"github.com/crunchydata/postgres-operator/events"
-	"github.com/crunchydata/postgres-operator/operator/patroni"
+	"github.com/crunchydata/postgres-operator/internal/patroni"
 	log "github.com/sirupsen/logrus"
 
 	kerrors "k8s.io/apimachinery/pkg/api/errors"
